requirements.txt
cycler==0.10.0
keyboard==0.13.5
kiwisolver==1.3.1
matplotlib==3.4.3
MouseInfo==0.1.3
numpy==1.21.2
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
PyTweening==1.0.3
six==1.16.0
